pyproject.toml
[tool.poetry]
name = "fobi-poc"
version = "0.1.0"
description = ""
authors = ["Bruno Alves <brunxbk@gmail.com>"]
readme = "README.md"
packages = [{include = "fobi_poc"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
anyascii = "0.3.2"
asgiref = "3.7.2"
asttokens = "2.4.1"
attrs = "23.1.0"
beautifulsoup4 = "4.11.2"
black = "22.3.0"
bleach = "6.1.0"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
curlylint = "0.13.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
dj-database-url = "0.4.1"
django-autoslug = "1.9.9"
django-cors-headers = "4.3.1"
django-csp = "3.7"
django-dotenv = "1.4.1"
django-extensions = "3.2.1"
django-filter = "23.4"
django-fobi = "0.19.9"
django-formtools = "2.4.1"
django-modelcluster = "6.1"
django-nine = "0.2.7"
django-nonefield = "0.4"
django-permissionedforms = "0.1"
django-taggit = "4.0.0"
django-treebeard = "4.7"
django-debug-toolbar = "3.8.1"
djangorestframework = "3.14.0"
djhtml = "1.4.13"
draftjs-exporter = "2.1.7"
easy-thumbnails = "2.8.5"
et-xmlfile = "1.1.0"
executing = "2.0.1"
filetype = "1.2.0"
flake8 = "6.1.0"
flake8-assertive = "2.0.0"
flake8-blind-except = "0.1.1"
flake8-comprehensions = "3.8.0"
flake8-print = "5.0.0"
html5lib = "1.1"
idna = "3.6"
ipdb = "0.13.13"
ipython = "8.18.1"
isort = "5.6.4"
jedi = "0.19.1"
l18n = "2021.3"
matplotlib-inline = "0.1.6"
mccabe = "0.7.0"
mypy-extensions = "1.0.0"
openpyxl = "3.1.2"
packaging = "23.2"
parso = "0.8.3"
parsy = "1.1.0"
pathspec = "0.11.2"
pexpect = "4.9.0"
pillow = "10.1.0"
pillow-heif = "0.13.1"
platformdirs = "4.0.0"
prompt-toolkit = "3.0.41"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycodestyle = "2.11.1"
pyflakes = "3.1.0"
pygments = "2.17.2"
pytz = "2023.3.post1"
requests = "2.31.0"
six = "1.16.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
stack-data = "0.6.3"
telepath = "0.3.1"
toml = "0.10.2"
traitlets = "5.14.0"
unidecode = "1.3.7"
urllib3 = "2.1.0"
vishap = "0.1.5"
wagtail = "5.2.1"
wagtail-font-awesome-svg = "0.0.4"
wcwidth = "0.2.12"
webencodings = "0.5.1"
willow = "1.6.3"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
